Updated LotteryResultsFunction to get wining numbers from new LotterResults dynamo table. Updated frontend to handle new results format.
