+++ backend/functions/lottery_results/requirements.txt
